feat: tarot-codex route by cardName, search bar, upright/reversed layout
- Change [id] to [cardName] (slug: the-fool, ace-of-wands)
- Add search bar to filter cards by name, keywords, meaning
- Beautify edit page: split upright/reversed sections
- Upright: image + upright fields; Reversed: rotated image + reversed fields
- Remove background from card image containers
